Give friendly animals a little nudge to move when you're trying to build
Only sheep and zebras move. Wolves don't care about you or your city.
Animals only move if they are alive and at rest. Animals which path into the area will redirect as soon as they come to rest and go elsewhere.
Animals try to move to a (more-or-less random) adjacent spot, if it is available to be walked on. If there are no available adjacent spots, they do nothing
Does not work for hippodrome, for whatever reason. We can tweak this later if it's a major concern.